+++ pkg/service/leases.go
@@ -104,47 +104,71 @@ func (s *LeasesService) ListLeases(acctID, principalID, nextAcctID, nextPrincipa
 	log.Infoln(string(jsonPayload))
 }
 
-func (s *LeasesService) LoginToLease(leaseID, loginProfile string, loginOpenBrowser, loginPrintCreds bool) {
+type leaseCreds struct {
+	AccessKeyID string `json:"accessKeyId,omitempty"`
+	ConsoleURL string `json:"consoleUrl,omitempty"`
+	ExpiresOn float64 `json:"expiresOn,omitempty"`
+	SecretAccessKey string `json:"secretAccessKey,omitempty"`
+	SessionToken string `json:"sessionToken,omitempty"`
+}
+
+func (s *LeasesService) Login(opts *LeaseLoginOptions) {
 	log.Debugln("Requesting leased account credentials")
-	params := &operations.PostLeasesIDAuthParams{
-		ID: leaseID,
-	}
-	params.SetTimeout(5 * time.Second)
-	res, err := apiClient.PostLeasesIDAuth(params, nil)
+
+	params := &operations.PostLeasesAuthParams{}
+	params.SetTimeout(20 * time.Second)
+	res, err := apiClient.PostLeasesAuth(params, nil)
+
 	if err != nil {
-		log.Fatalln("err: ", err)
-	} else {
-		jsonPayload, err := json.MarshalIndent(res.GetPayload(), "", "\t")
-		if err != nil {
-			log.Fatalln("err: ", err)
-		}
-		log.Debug(string(jsonPayload))
+		log.Fatal(err)
 	}
 
 	responsePayload := res.GetPayload()
 
-	if !(loginOpenBrowser || loginPrintCreds) {
+	creds := leaseCreds(*responsePayload)
+	s.loginWithCreds(&creds, opts)
+}
+
+func (s *LeasesService) LoginByID(leaseID string, opts *LeaseLoginOptions) {
+	log.Debugln("Requesting leased account credentials")
+	params := &operations.PostLeasesIDAuthParams{
+		ID:         leaseID,
+	}
+	params.SetTimeout(20 * time.Second)
+	res, err := apiClient.PostLeasesIDAuth(params, nil)
+	if err != nil {
+		log.Fatalln("err: ", err)
+	}
+
+	responsePayload := res.GetPayload()
+
+	creds := leaseCreds(*responsePayload)
+	s.loginWithCreds(&creds, opts)
+}
+
+func (s *LeasesService) loginWithCreds(leaseCreds *leaseCreds, opts *LeaseLoginOptions) {
+	if !(opts.OpenBrowser || opts.PrintCreds) {
 		credsPath := filepath.Join(".aws", "credentials")
 		log.Infoln("Adding credentials to " + credsPath + " using AWS CLI")
-		s.Util.ConfigureAWSCLICredentials(responsePayload.AccessKeyID,
-			responsePayload.SecretAccessKey,
-			responsePayload.SessionToken,
-			loginProfile)
+		s.Util.ConfigureAWSCLICredentials(leaseCreds.AccessKeyID,
+			leaseCreds.SecretAccessKey,
+			leaseCreds.SessionToken,
+			opts.CliProfile)
 
-	} else if loginProfile != "default" {
+	} else if opts.CliProfile != "default" {
 		log.Infoln("Setting --profile has no effect when used with other flags.\n")
 	}
 
-	if loginOpenBrowser {
+	if opts.OpenBrowser {
 		log.Infoln("Opening AWS Console in Web Browser")
-		s.Util.OpenURL(responsePayload.ConsoleURL)
+		s.Util.OpenURL(leaseCreds.ConsoleURL)
 	}
 
-	if loginPrintCreds {
+	if opts.PrintCreds {
 		creds := fmt.Sprintf(constants.CredentialsExport,
-			responsePayload.AccessKeyID,
-			responsePayload.SecretAccessKey,
-			responsePayload.SessionToken)
+			leaseCreds.AccessKeyID,
+			leaseCreds.SecretAccessKey,
+			leaseCreds.SessionToken)
 		log.Infoln(creds)
 	}
 }