+++ pkg/service/service.go
@@ -72,10 +72,17 @@ type Accounter interface {
 	ListAccounts()
 }
 
+type LeaseLoginOptions struct {
+	CliProfile  string
+	OpenBrowser bool
+	PrintCreds  bool
+}
+
 type Leaser interface {
 	CreateLease(principalID string, budgetAmount float64, budgetCurrency string, email []string, expiresOn string)
 	EndLease(accountID, principalID string)
-	LoginToLease(leaseID, profile string, loginOpenBrowser, loginPrintCreds bool)
+	LoginByID(leaseID string, opts *LeaseLoginOptions)
+	Login(opts *LeaseLoginOptions)
 	ListLeases(acctID, principalID, nextAcctID, nextPrincipalID, leaseStatus string, pagLimit int64)
 	GetLease(leaseID string)
 }